+++ src/components/app-header.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const AppHeader = () => {
+  return <h1>My Todo List</h1>;
+};
+
+export default AppHeader;
