+++ src/components/search-panel.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SearchPanel = () => {
+  return <input placeholder="search" />;
+};
+
+export default SearchPanel
